A chatbot is a computer program designed to simulate human conversation.Chatbot using PHP with MySQL & jQuery (Ajax)
 These chatbots reply to you instantly according to your queries because programmers have inserted thousands of inputs/replies/queries into the database that can be asked by the user. To make an advanced chatbot we've to code more and more but I tried to make a simple chatbot with few lines of codes and queries which help you to get an idea about how chatbot actually works.

In this program [Simple Chatbot using PHP], on the webpage, there is a chat form with an input field and a button labeled as "send" for typing a message and send to the bot. When you ask something to the bot and that query exists in the database then the bot replay a message according to your query instantly but if your query not matched to the database queries then the bot replays a message labeled as "Sorry can't be able to understand you!". In this chatting process, the webpage isn't reloaded because I used jQuery (Ajax) for that.